1。添加mysql支持 2，ef sql日志 3，添加redis 支持 4，添加加密工具 5，添加ef 扩展工具 6，其他代码优化

+++ Hei.Admin.Repository/Migrations/HuachDbContextModelSnapshot.cs
@@ -437,7 +437,7 @@ namespace Hei.Admin.Repository.Migrations
                         .IsRequired()
                         .HasMaxLength(20);
 
-                    b.Property<int>("Type");
+                    b.Property<short>("Type");
 
                     b.Property<string>("UserImg")
                         .IsRequired()
